fix: 🐛 Typescript version mismatch causing issues with Intellisense (#332)
We were getting a weird infinite "Loading..." issue with VS Code Typescript Intellisense. We determined this was caused by a mismatch with the update of the Typescript version in #221 with the auth changes. 

We determined either downgrading the Typescript version to the one prior to the update in #221 fixed the issue, or removing the `node_modules` folder, along with the `pnpm-lock.yaml` file and running a fresh `pnpm` install.